Исправил автообновление главной формы. Добавил измерение времени занятого для ответа. НУЖНО подумать как реализовать правку преференсов "не полных" тестов. (думаю может отправить их на новую наследуемую страницу)

+++ Models/Quest.cs
@@ -22,7 +22,7 @@ namespace TestMulti.Models
 
         public string QuestColor { get; set; } = "Gray"; // Серый цвет по умолчанию
         public string AnswerColor { get; set; } = "Gray";  // Серый цвет по умолчанию
-        public int Ellapsed { get; set; } = 0; // Время, прошедшее с момента начала вопроса
+        public double Ellapsed { get; set; } = 0; // Время, прошедшее с момента начала вопроса
 
         public static Quest [] GetForLearn(Quest[] quests)
         {
